+++ EnableGPlayWithPC/EnableGPlayWithPC.csproj
@@ -42,7 +42,8 @@
       <HintPath>..\packages\WindowsAPICodePack-Shell.1.1.1\lib\Microsoft.WindowsAPICodePack.Shell.dll</HintPath>
     </Reference>
     <Reference Include="SharpAdbClient, Version=2.2.0.0, Culture=neutral, PublicKeyToken=d728076f5d82a2e8, processorArchitecture=MSIL">
-      <HintPath>..\packages\SharpAdbClient.2.2.35\lib\net451\SharpAdbClient.dll</HintPath>
+      <SpecificVersion>False</SpecificVersion>
+      <HintPath>C:\lib\SharpAdbClient.dll</HintPath>
     </Reference>
     <Reference Include="System" />
     <Reference Include="System.Core" />
@@ -58,7 +59,7 @@
   <ItemGroup>
     <Compile Include="Apks.cs" />
     <Compile Include="BenesseTabs.cs" />
-    <Compile Include="ErrorDialog.cs" />
+    <Compile Include="Dialog.cs" />
     <Compile Include="FileSelector.cs">
       <SubType>UserControl</SubType>
     </Compile>
